Move Shader from pants.java to the jvm backend.
As things stood, the core pants.java utilties had to depend on the jvm
backend, forcing backends like python to have to do said same
transitively.  All uses os the Shader are in the jvm backend, so the
move is clean.
