+++ src/python/pants/backend/jvm/tasks/BUILD
@@ -67,14 +67,14 @@ python_library(
   dependencies = [
     ':ivy_task_mixin',
     ':jar_task',
+    'src/python/pants/backend/jvm/subsystems:jvm_tool_mixin',
+    'src/python/pants/backend/jvm/subsystems:shader',
     'src/python/pants/base:cache_manager',
     'src/python/pants/base:exceptions',
     'src/python/pants/base:workunit',
     'src/python/pants/ivy',
     'src/python/pants/java:executor',
     'src/python/pants/java:util',
-    'src/python/pants/java/jar:shader',
-    'src/python/pants/backend/jvm/subsystems:jvm_tool_mixin',
     'src/python/pants/util:dirutil',
     'src/python/pants/util:memo',
   ],
@@ -110,9 +110,9 @@ python_library(
   dependencies = [
     '3rdparty/python/twitter/commons:twitter.common.collections',
     ':nailgun_task',
+    'src/python/pants/backend/jvm/subsystems:shader',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/base:exceptions',
-    'src/python/pants/java/jar:shader',
     'src/python/pants/option',
     'src/python/pants/process',
     'src/python/pants/util:dirutil',
@@ -274,12 +274,12 @@ python_library(
     ':jvm_task',
     ':jvm_tool_task_mixin',
     'src/python/pants/backend/core/tasks:task',
+    'src/python/pants/backend/jvm/subsystems:shader',
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:workunit',
     'src/python/pants/binaries:binary_util',
-    'src/python/pants/java/jar:shader',
     'src/python/pants/java:util',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
@@ -291,11 +291,11 @@ python_library(
   name = 'jvm_binary_task',
   sources = ['jvm_binary_task.py'],
   dependencies = [
-    ':jar_task',
     '3rdparty/python/twitter/commons:twitter.common.collections',
+    ':jar_task',
+    'src/python/pants/backend/jvm/subsystems:shader',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/base:exceptions',
-    'src/python/pants/java/jar:shader',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:fileutil',
     'src/python/pants/util:memo',
