+++ src/python/pants/backend/jvm/tasks/jvm_binary_task.py
@@ -11,10 +11,10 @@ from contextlib import contextmanager
 
 from twitter.common.collections.orderedset import OrderedSet
 
+from pants.backend.jvm.subsystems.shader import Shader
 from pants.backend.jvm.targets.jvm_binary import JvmBinary
 from pants.backend.jvm.tasks.jar_task import JarTask
 from pants.base.exceptions import TaskError
-from pants.java.jar.shader import Shader
 from pants.java.util import execute_runner
 from pants.util.contextutil import temporary_dir
 from pants.util.fileutil import atomic_copy
