+++ src/python/pants/backend/jvm/tasks/jvm_compile/BUILD
@@ -159,12 +159,12 @@ python_library(
     ':analysis_tools',
     ':jvm_compile',
     'src/python/pants/backend/jvm/subsystems:scala_platform',
+    'src/python/pants/backend/jvm/subsystems:shader',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/backend/jvm/targets:scala',
     'src/python/pants/base:exceptions',
     'src/python/pants/base:hash_utils',
     'src/python/pants/base:workunit',
-    'src/python/pants/java/jar:shader',
     'src/python/pants/option',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
