+++ tests/python/pants_test/backend/jvm/subsystems/test_shader_integration.py
