+++ tests/python/pants_test/java/jar/BUILD
@@ -5,7 +5,6 @@ target(
   name = 'jar',
   dependencies = [
     ':manifest',
-    ':shader'
   ]
 )
 
@@ -15,30 +14,4 @@ python_tests(
   dependencies = [
     'src/python/pants/java/jar:manifest',
   ]
-)
-
-python_tests(
-  name = 'shader',
-  sources = ['test_shader.py'],
-  dependencies = [
-    'src/python/pants/java/distribution',
-    'src/python/pants/java/jar:shader',
-    'src/python/pants/java:executor',
-    'src/python/pants/util:contextutil',
-    'src/python/pants/util:dirutil',
-    'tests/python/pants_test/subsystem:subsystem_utils',
-  ]
-)
-
-python_tests(
-  name='shader_integration',
-  sources=['test_shader_integration.py'],
-  dependencies=[
-    'src/python/pants/fs',
-    'src/python/pants/java/distribution',
-    'src/python/pants/java:executor',
-    'src/python/pants/util:contextutil',
-    'tests/python/pants_test/subsystem:subsystem_utils',
-    'tests/python/pants_test:int-test',
-  ],
 )
